FixSchedual

+++ src/main/java/tn/esprit/bank/vo/TransactionVO.java
@@ -3,12 +3,13 @@ package tn.esprit.bank.vo;
 import lombok.Data;
 
 import java.math.BigDecimal;
+import java.math.BigInteger;
 
 @Data
 public class TransactionVO {
 
-    private BigDecimal accountNumberFrom;
-    private BigDecimal accountNumberTo;
+    private BigInteger accountNumberFrom;
+    private BigInteger accountNumberTo;
     private BigDecimal amount;
